+++ .github/workflows/main.yaml
@@ -8,6 +8,7 @@ on:
   push:
     branches:
       - main
+      - dev
   pull_request:
     branches:
       - main
@@ -28,38 +29,26 @@ jobs:
         **/*.md
       modified_files_only: false
 
-#  flutter_test:
-#    name: Run Flutter Test And Analyze
-#    runs-on: ubuntu-latest
-#    steps:
-#      - uses: actions/checkout@v1
-#      - uses: subosito/flutter-action@v1
-#        with:
-#          flutter-version: '2.5.0'
-#          channel: 'stable'
-#      - run: flutter pub get
-#      - run: flutter test
-
-#  build_appbundle:
-#    name: Build Flutter (Android)
-#    needs: [flutter_test]
-#    runs-on: ubuntu-latest
-#    steps:
-#      - uses: actions/checkout@v2
-#      - uses: actions/setup-java@v2
-#        with:
-#          distribution: 'zulu'
-#          java-version: '11'
-#      - uses: subosito/flutter-action@v2
-#        with:
-#          channel: "stable"
-#      - run: flutter pub get
-#      - run: flutter clean
-#      - run: flutter build apk --debug --split-per-abi
-#      - name: Push to Releases
-#        uses: ncipollo/release-action@v1
-#        with:
-#          artifacts: "build/app/outputs/apk/debug/*,build/ios/iphoneos/app.ipa"
-#          tag: v1.0.${{ github.run_number }}
-#          token: ${{ secrets.TOKEN }}
+  build_apk:
+    name: Build OnSpace APK (Production)
+    needs: [semantic-pull-request, build, spell-check]
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-java@v2
+        with:
+          distribution: 'zulu'
+          java-version: '11'
+      - uses: subosito/flutter-action@v2
+        with:
+          channel: "stable"
+      - run: flutter pub get
+      - run: flutter clean
+      - run: flutter build apk --flavor=production -t lib/main_production.dart
+      - name: Push to Releases
+        uses: ncipollo/release-action@v1
+        with:
+          artifacts: "build/app/outputs/flutter-apk/*"
+          tag: v1.0.${{ github.run_number }}
+          token: ${{ secrets.TOKEN }}
